% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_dangerous_calls.R
\name{default_dangerous_calls}
\alias{default_dangerous_calls}
\title{Default List of Dangerous Functions}
\usage{
default_dangerous_calls()
}
\value{
A character vector of default forbidden function names.
}
\description{
Returns the default set of dangerous function names that are checked
by \code{\link[=find_dangerous_calls]{find_dangerous_calls()}}.
}
