% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_doc_prompt.R
\name{ggplot2_doc_prompt}
\alias{ggplot2_doc_prompt}
\title{Create an LLM system prompt using ggplot2 Documentation}
\usage{
ggplot2_doc_prompt()
}
\value{
an \code{ellmer_prompt} object with the system prompt suitable for use creating
an \code{ellmer::Chat} object.
}
\description{
This function creates a system prompt for an LLM that includes the contents of help pages
for 'ggplot2' functions related to themes.  This is useful for creating
\code{\link[ellmer:Chat]{ellmer::Chat}} objects that are fully aware of the documented state of the API.
}
