% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_news_prompt.R
\name{ggplot2_news_prompt}
\alias{ggplot2_news_prompt}
\title{Create an LLM system prompt using ggplot2 NEWS}
\usage{
ggplot2_news_prompt()
}
\value{
an \code{ellmer_prompt} object with the system prompt suitable for use creating
an \code{ellmer::Chat} object.
}
\description{
This function creates a system prompt for an LLM that includes the contents of the
news for the latest major version of \strong{ggplot2}. This is useful for
creating \code{\link[ellmer:Chat]{ellmer::Chat}} objects that are aware of recent changes in the package.
However, doing this will incur extra costs from the LLM provider (if being billed per-token).
}
