% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_list_fonts.R
\name{tool_list_fonts}
\alias{tool_list_fonts}
\title{Define a tool to list available fonts}
\usage{
tool_list_fonts()
}
\value{
An S7 \code{ToolDef} object.
}
\description{
Defines a tool that lists all available font families on the system for
a chatbot. This is intended to be used with an \code{ellmer::Chat} object.
}
\examples{
\dontrun{
chat <- ellmer::chat_openai()

chat$register_tool(tool_list_fonts)

chat$chat('What font families are on this system')

}
}
